% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofit_gaussian_2D.R
\name{autofit_gaussian_2D}
\alias{autofit_gaussian_2D}
\title{Automatically determine the best-fitting 2D Gaussian for a data set}
\usage{
autofit_gaussian_2D(data, comparison_method = "rmse", maxiter = 1000)
}
\arguments{
\item{data}{A data.frame that contains the raw data (generally rectilinearly
gridded data, but this is not a strict requirement). Columns must be named
\code{"X_values"}, \code{"Y_values"} and \code{"response"}.}

\item{comparison_method}{One of "rmse", "rss", or "AIC"; what metric should
be used to determine the "best-fitting" Gaussian?}

\item{maxiter}{Default 1000. A positive integer specifying the maximum number
of iterations allowed. See \code{stats::nls.control()} for more details.}
}
\value{
A list with the components:
\itemize{
\item{"coefs"} {A data.frame of fitted model parameters.}
\item{"model"} {The model object, fitted by \code{stats::nls()}.}
\item{"model_error_stats"} {A data.frame detailing the rss, rmse, deviance,
and AIC of the fitted model.}
\item{"fit_method"} {A character vector that indicates which method and
orientation strategy was used by this function.}
}
}
\description{
Automatically determine the best-fitting 2D Gaussian for a data set
}
\details{
This function runs \code{fit_gaussian_2D()} three times: once for
each of the "main" types of models: 1) elliptical, unconstrained; 2)
elliptical, log; 3) circular. The function \code{compare_gaussian_fits()} is
then used to determine which of these three models is the best-fitting, using
the \code{comparison_method} argument to make the decision.
}
\examples{
if (interactive()) {
  library(gaussplotR)

  ## Load the sample data set
  data(gaussplot_sample_data)

  ## The raw data we'd like to use are in columns 1:3
  samp_dat <-
    gaussplot_sample_data[,1:3]

  ## Automatically determine the best model to use
  gauss_auto <-
    autofit_gaussian_2D(samp_dat)
}
}
\author{
Vikram B. Baliga
}
