% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_gaussian_2D.R
\name{ggplot_gaussian_2D}
\alias{ggplot_gaussian_2D}
\title{Plot a 2D gaussian via ggplot}
\usage{
ggplot_gaussian_2D(
  gauss_data,
  normalize = TRUE,
  contour_thickness = 0.04,
  bins = 15,
  viridis_dir = 1,
  viridis_opt = "B",
  x_lab = "X values",
  y_lab = "Y values",
  axis.text = element_text(size = 6),
  axis.title = element_text(size = 7),
  axis.ticks = element_line(size = 0.3),
  plot.margin = unit(c(0.1, 0.1, 0.1, 0.1), "cm")
)
}
\arguments{
\item{gauss_data}{Data.frame with X_values, Y_values, and predicted_values,
e.g. exported from \code{predict_gaussian_2D()}}

\item{normalize}{Default TRUE, should predicted_values be normalized on a 0
to 1 scale?}

\item{contour_thickness}{Thickness of contour lines}

\item{bins}{Number of bins for the contour plot}

\item{viridis_dir}{See "direction" in scale_fill_viridis_c()}

\item{viridis_opt}{See "option" in scale_fill_viridis_c()}

\item{x_lab}{Arguments passed to xlab()}

\item{y_lab}{Arguments passed to ylab()}

\item{axis.text}{Arguments passed to axis.text}

\item{axis.title}{Arguments passed to axis.title}

\item{axis.ticks}{Arguments passed to axis.ticks}

\item{plot.margin}{Arguments passed to plot.margin}
}
\value{
A ggplot object that uses metR::geom_contour_fill() to display the
2D gaussian
}
\description{
Plot a 2D gaussian via ggplot
}
\examples{
if (interactive()) {
  ## Load the sample data set
  data(gaussplot_sample_data)

  ## The raw data we'd like to use are in columns 1:3
  samp_dat <-
    gaussplot_sample_data[,1:3]


  #### Example 1: Unconstrained elliptical ####
  gauss_fit <-
    fit_gaussian_2D(samp_dat)
  ## This fits an unconstrained elliptical by default
  attr(gauss_fit, "fit_method")

  ## Generate a grid of x- and y- values on which to predict
  grid <-
    expand.grid(X_values = seq(from = -5, to = 0, by = 0.1),
                Y_values = seq(from = -1, to = 4, by = 0.1))

  ## Predict the values using predict_gaussian_2D
  gauss_data <-
    predict_gaussian_2D(
      fit_params = gauss_fit,
      X_values = grid$X_values,
      Y_values = grid$Y_values,
    )

  ## Plot via ggplot2 and metR
  library(ggplot2); library(metR)
  ggplot_gaussian_2D(gauss_data)

  #### Example 2: Constrained elliptical_log ####
  gauss_fit <-
    fit_gaussian_2D(
      samp_dat,
      method = "elliptical_log",
      orientation_strategy = -1
    )
  ## This fits a constrained elliptical, as in Priebe et al. 2003
  attr(gauss_fit, "fit_method")

  ## Generate a grid of x- and y- values on which to predict
  grid <-
    expand.grid(X_values = seq(from = -5, to = 0, by = 0.1),
                Y_values = seq(from = -1, to = 4, by = 0.1))

  ## Predict the values using predict_gaussian_2D
  gauss_data <-
    predict_gaussian_2D(
      fit_params = gauss_fit,
      X_values = grid$X_values,
      Y_values = grid$Y_values,
    )

  ## Plot via ggplot2 and metR
  ggplot_gaussian_2D(gauss_data)
}
}
\author{
Vikram B. Baliga
}
