% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gaussian_2D.R
\name{predict_gaussian_2D}
\alias{predict_gaussian_2D}
\title{Predict values from a fitted 2D gaussian}
\usage{
predict_gaussian_2D(X_values, Y_values, A, X_peak, X_var, Q, Y_peak, Y_var)
}
\arguments{
\item{X_values}{vector of numeric values for the x-axis}

\item{Y_values}{vector of numeric values for the y-axis}

\item{A}{numeric value of the amplitude}

\item{X_peak}{numeric value of the x-location of the peak of the gaussian}

\item{X_var}{numeric value of the x-axis variance}

\item{Q}{numeric value for the orientation parameter}

\item{Y_peak}{numeric value of the y-location of the peak of the gaussian}

\item{Y_var}{numeric value of the y-axis variance}
}
\value{
A numeric vector that provides the values of the 2D gaussian,
corresponding to the set of {\code{X_values},\code{Y_values}}
}
\description{
Predict values from a fitted 2D gaussian
}
\section{Warning}{

This function assumes raw data have been log2 transformed
}

\author{
Vikram B. Baliga
}
